requirements.txt
datasets==2.16.1
huggingface-hub==0.20.2
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
streamlit==1.29.0
transformers==4.36.2
locust==2.23.1
click==8.0.3
tqdm==4.62.3
PyYAML==6.0